+++ src/schema/subscription/applicationEdited.subscription.ts
