+++ src/schema/subscription/index.ts
@@ -1,8 +1,8 @@
 import { GraphQLObjectType } from 'graphql';
-import applicationUnlocked from './applicationUnlocked';
-import applicationEdited from './applicationEdited';
-import notification from './notification';
-import recordAdded from './recordAdded';
+import applicationUnlocked from './applicationUnlocked.subscription';
+import applicationEdited from './applicationEdited.subscription';
+import notification from './notification.subscription';
+import recordAdded from './recordAdded.subscription';
 
 /** GraphQL subscriptions type definition */
 const Subscription = new GraphQLObjectType({
